+++ gazebo/msgs/msgs.cc
@@ -17,6 +17,7 @@
 
 #include <google/protobuf/descriptor.h>
 #include <algorithm>
+#include <array>
 #include <ignition/math/MassMatrix3.hh>
 #include <ignition/math/Rand.hh>
 
@@ -1807,7 +1808,8 @@ namespace gazebo
       if (_sdf->HasElement("topic"))
         result.set_topic(_sdf->Get<std::string>("topic"));
 
-      if (type == "camera" || type == "depth")
+      if (type == "camera" || type == "depth"
+          || type == "wideanglecamera" || type == "wide_angle_camera")
       {
         result.mutable_camera()->CopyFrom(
             msgs::CameraSensorFromSDF(_sdf->GetElement("camera")));