+++ gazebo/rendering/CameraPrivate.hh
@@ -117,6 +117,12 @@ namespace gazebo
 
       /// \brief Fixed axis to yaw around.
       public: ignition::math::Vector3d yawFixedAxis;
+
+      /// \brief Anti-aliasing value
+      public: uint32_t antiAliasingValue;
+
+      /// \brief Camera Intrinsic Matrix
+      public: ignition::math::Matrix3d cameraIntrinsicMatrix;
     };
   }
 }