+++ plugins/LensFlareSensorPlugin.cc
@@ -33,12 +33,18 @@ namespace gazebo
     /// \brief Lens flare
     public: std::vector<rendering::LensFlarePtr> lensFlares;
 
+    /// \brief Name of light that generates lens flare
+    public: std::string lightName;
+
     /// \brief Lens flare scale
     public: double scale = 1.0;
 
     /// \brief Lens flare color
     public: ignition::math::Vector3d color
         = ignition::math::Vector3d(1.4, 1.2, 1.0);
+
+    /// \brief Lens flare occlusion steps
+    public: double occlusionSteps = 10.0;
 
     /// \brief Lens flare compositor name
     public: std::string compositorName;
@@ -75,6 +81,11 @@ void LensFlareSensorPlugin::Load(sensors::SensorPtr _sensor,
     return;
   }
 
+  if (_sdf->HasElement("light_name"))
+  {
+    this->dataPtr->lightName = _sdf->Get<std::string>("light_name");
+  }
+
   if (_sdf->HasElement("scale"))
   {
     this->dataPtr->scale = _sdf->Get<double>("scale");
@@ -85,6 +96,11 @@ void LensFlareSensorPlugin::Load(sensors::SensorPtr _sensor,
   if (_sdf->HasElement("color"))
   {
     this->dataPtr->color = _sdf->Get<ignition::math::Vector3d>("color");
+  }
+
+  if (_sdf->HasElement("occlusion_steps"))
+  {
+    this->dataPtr->occlusionSteps = _sdf->Get<double>("occlusion_steps");
   }
 
   const std::string compositorName = "compositor";
@@ -120,6 +136,17 @@ void LensFlareSensorPlugin::Load(sensors::SensorPtr _sensor,
 }
 
 /////////////////////////////////////////////////
+void LensFlareSensorPlugin::SetLightName(std::string _name)
+{
+  this->dataPtr->lightName = _name;
+
+  for (auto flare : this->dataPtr->lensFlares)
+  {
+    flare->SetLightName(_name);
+  }
+}
+
+/////////////////////////////////////////////////
 void LensFlareSensorPlugin::SetScale(const double _scale)
 {
   this->dataPtr->scale = _scale;
@@ -142,6 +169,17 @@ void LensFlareSensorPlugin::SetColor(const ignition::math::Vector3d &_color)
 }
 
 /////////////////////////////////////////////////
+void LensFlareSensorPlugin::SetOcclusionSteps(double _occlusionSteps)
+{
+  this->dataPtr->occlusionSteps = _occlusionSteps;
+
+  for (auto flare : this->dataPtr->lensFlares)
+  {
+    flare->SetOcclusionSteps(_occlusionSteps);
+  }
+}
+
+/////////////////////////////////////////////////
 void LensFlareSensorPlugin::AddLensFlare(rendering::CameraPtr _camera)
 {
   if (!_camera)
@@ -153,8 +191,11 @@ void LensFlareSensorPlugin::AddLensFlare(rendering::CameraPtr _camera)
   {
     lensFlare->SetCompositorName(this->dataPtr->compositorName);
   }
-  lensFlare->SetCamera(_camera);
+  lensFlare->SetLightName(this->dataPtr->lightName);
   lensFlare->SetScale(this->dataPtr->scale);
   lensFlare->SetColor(this->dataPtr->color);
+  lensFlare->SetOcclusionSteps(this->dataPtr->occlusionSteps);
+  // SetCamera must be called last
+  lensFlare->SetCamera(_camera);
   this->dataPtr->lensFlares.push_back(lensFlare);
 }